webhooks/requirements.txt
discord
discord-interactions
flask[async]
python-dateutil
python-dotenv
pytz
requests
